Extend DataType fungibility to include integer-type conversions
DataTypes are intended to be fungible, meaning that Int(10)==BigInt(10),
but in the current implementation this does not extend to integer type
conversions -- that is, i32::from(Int(10)) == 10, but i32::from(BigInt(10))
panics. This patch allows for conversions from any integer datatype to any
integer value that fits the contents of that datatype. With this patch,
converting from an int now also always converts into the DataType with the
smallest MAX that contains the input -- so, DataType::From(10) will result
in a DataType::Int.