Fix nested table name rewrites
When noria is renaming table aliases in queries, it does not recurse
into expressions. For instance, it will rename
`SELECT foo.bar FROM table as foo` to `SELECT table.bar FROM table`
but will not perform any renaming on:
`SELECT count(foo.bar) FROM table as foo`

This patch recurses into expressions to add this renaming in
expressions, along with a limited test, testing two instances of
this functionality (by analogy with the existing test).