Add a lookup_range method to the State trait
Add a method to the State trait, which represents the backing storage
for internal and base table nodes (but not reader nodes), which allows
looking up *ranges* of keys in addition to just exact keys, similar to
what we did for evbtree in readers. To support this, the
differently-sized variants of the KeyedState type now contain a BTreeMap
instead of a HashMap, and we had to do the same thing here that we did
with the reader where we made eviction walk the list of keys to pick
random ones (ugh) instead of evicting by index. Fortunately, rocksdb
supports ordered range iteration natively and performantly, so doing
this there was quite easy.