Migrate from the failure crate to anyhow + thiserror
Rust's error handling story is a bit broken, partially due to the flawed
`std::error::Error` trait in the standard library. When `failure` was created,
this trait was even more useless than it currently is, leading to that crate
being introduced and defining a new `Fail` trait (plus easy ways to derive it).
Now that `std::error::Error` has gained some of the capabilities of `Fail`,
`failure` is considered deprecated, and `anyhow` is somewhat of a drop-in
replacement that has a similar API, but uses the stdlib's trait.

One piece of functionality `failure` provided that `anyhow` does not is the
ability to deal with `Box<dyn std::error::Error>` types. These suck a whole
bunch, because the standard library is missing an implementation of `Error`
for them (recall that boxed trait objects don't implement the trait
automatically!) -- and we can't avoid dealing with these trait objects because
the `tower_buffer` crate noria uses creates them (because urgh futures
and associated types are hard). This is the first sucky thing we have to deal
with.

To convert this into something useful, we use a newtype cribbed from the
`failure` crate. In an ideal world, though, we'd just downcast the boxed trait
object into the error it actually is -- but we can't do that here because the
error a lot of the tower services generate is `anyhow::Error`, which doesn't
actually implement `std::error::Error` (!!!). This is the second sucky thing
we have to deal with (but we can actually hope to avoid this in future, by
being more careful with error handling generally and making it into a more
specific enum).

Oh, and the `thiserror` crate is now used to fill in for the removed `failure`'s
derive capabilities (which is relatively uncontroversial).