Implement deletion in unbounded-interval-tree
Implement unbounded_interval_tree::IntervalTree::remove, to remove a
given interval from the tree. This is surprisingly complex and hard to
get right -see inline comments in the function for a general explanation
of how this works. To aid in making sure it *does* work with all the
edge cases and subtleties, I've introduced `proptest` to this crate with
an implementation of `Arbitrary` for IntervalTree, and removal is
covered by two property tests in addition to a couple of example tests -
one that checks that once we remove an interval it isn't there any more,
and another that checks that removing an interval preserves other
intervals disjoint to the one we're removing. Also, all of these tests
now validate the invariants of the tree after every operation.

Another thing worth talking about in this commit is how annoying the
Bound type is to deal with in Rust - most notably, since it doesn't
differentiate between positive unbounded and negative unbounded, we have
to have *four* different comparison functions for Bound rather than just
implementing Ord - one for comparing each of the posssible combinations
of start bound and end bound. These, plus predicates to test whether
intervals overlap or completely cover each other, and the ordering
relation on intervals we use for actually sorting the interval tree
itself, have been extracted to a new private `intervals` module with
extensive property tests of its own.