Add support for avg operator as a new variant of numerical
The average operator relies on additional, non-column state.
Specifically a HashMap to maintain sum and count for each group from
which to calculate the average.

Also considered implementing avg operator via a multiaggregate operator
that had columns for sum and count. Then, average could be computed
via a filter/projection node that divides sum by count. We went with
the simpler approach of implementing avg as a variant of a numerical
aggregator with additional state because the overhead of implementing
an entirely new multiaggregate operator was not deemed to be worth it
for this narrow use case. This implementation most closely matches
the existing methods for dealing with aggregations and was the quickest
way to obtain a working average operator.

If need for a multiaggregator arises in the future, aggregations could
be refactored in their entirety to support this and  average could be
changed to be implemented as a downstream projection off of a
multiaggregator.

Also considered implementing average as a join on separate sum and count
nodes. This was not selected due to added complexity of making sure the
separate sum and count operators are in sync when computing average.

Add support for  numerical aggregators to be over both integer and
decimal columns, and counts can be over any column.