Make node State range-aware
Update the State trait, which backs state for all internal nodes and
base tables (but not readers), to be aware of ranges in the same
way that readers now are. The trait has been updated so that
`mark_filled` now takes a `KeyComparison` (which can be a range) instead
of just a key, and the `MemoryState` has been updated to track which
ranges it has filled. Since base tables are never partial, there wasn't
much to do there other than just updating the types.

To support making MemoryState partial, I've encapsulated the notion of a
BTreeMap-with-interval-tree into a new PartialMap struct, which
implements the subset of the BTreeMap API that was necessary to make
this work. That's now the backing storage of the variants of
`KeyedState`.