Don't try to streaming-serialize range results
Apparently, the bincode serialization format needs to know the length of
sequences before beginning to serialize them - this is a thing that
serde lets particular serialization formats do, and they're expected to
just panic if a caller calls `serialize_seq(None)`. The alternative here
(unfortunately) is to clone the records into a vec before serializing
them all at once. This commit does that, and also covers it with an
integration test, which I really should've done in the first place