Add noria-mysql frontend for range queries
In order to support ranged queries in the noria-mysql frontend, we use the
stored SQL AST in do_read(), walking it to determine which binary operators were
used in the query (i.e. no longer assuming they're all just `Equal`). These are
then used to make a KeyComparison to pass to multi_lookup.

We modify get_parameter_columns_recurse() in utils.rs to return both the column
name and the operator used, so we can walk the AST later as described above.
(The new flip_comparison() method is used to handle cases where the literal
comes before the column name, i.e. "WHERE 3 < column".)