Add a remove_range method to evbtree
Add a method to evbtree to remove all the keys covered by a given range,
and add this as an operation to test in the quickcheck tests.

Unfortunately, since the rust btree library is missing a drain with
range method, this has to iterate the entire set of keys rather than
being able to use the structure of the btree to get an O(log(n))
removal. This is really upsetting, but will have to suffice for now
until we can get around to digging into the rust btree internals to get
something better.