Make all upqueries use a KeyComparsion
Update all code paths for upqueries to use a KeyComparison instead of
just a Vec<DataType>. This includes the upqueries themselves, the keys
associated with the snapshots in response, the internal state of a
domain containing the set of pending upqueries, the keys of a
BlockingRead, all eviction methods, the Misses, Lookups, and Captured
returned by an Ingredient in the NodeProcessingResult, and mark_hole and
mark_filled in both internal- and reader-node state. With this change,
we can now successfully query into partially materialized nodes with
ranges - and there's an integration test asserting exactly that.

The only thing that's not supported here is unions - there's some
interesting subtlety around applying deltas to buffered replays that
gets kinda complicated. Also, to make all of this *fully* work we need
to have logic for deduplicating ranged upqueries by overlaps rather than
just covers, which is omitted here.

There are a couple of minor refactors in this commit that aren't really
worth noting here, but one that *is* worth noting is that the
`WriteHandle` for a backlog now has a `read` method that returns a
ReadHandle, rather than duplicating all the various read
methods (try_find_and, contains_key, etc) in both the read- and
write-handle. This *seems* to maintain the left-right invariants and
remain internally consistent, but there could be something there I'm
misunderstanding that we could stand to look closer at in the future.