Properly copy over changes to the tree
There's an optimization on the write handle of a left-right where, if
publish() has never been called before, we apply all changes *directly*
to the underlying map rather than putting them in the oplog, and then
put an op in the log called JustCloneRHandle to copy all those chanages
over to the other side when we eventually *do* call publish. Previously,
the implementation of JustCloneRHandle was not properly copying over the
contents of the tree to the other side, meaning that calls to
`insert_range` were only seen by reads performed after an *odd* number
of publishes (aka only on one side of the left-right). This fixes that
by just cloning over the entire interval tree on calls to
JustCloneRHandle.