Add parser for sqllogictest test scripts
Add the beginning of a simple nom-based parser for the test script
format used by SQLite's sqllogictest[1], which is a simple verification
program for the correctness of SQL databases which the SQLite project
has a number of useful test scripts associated with. The idea behind
porting it to rust is that we can now build tooling around the existing
format, including converting each individual test to a Rust test case
using a proc macro, enhancing the test format with noria-specific
semantics (such as reordering non-side-effecting queries, etc.), etc.

[1]: https://www.sqlite.org/sqllogictest/doc/trunk/about.wiki