Make executing SELECT statements actually work
The new `absurdly_simple_select` test, while it might look like something you'd
hope would work, was horribly broken for 2 different reasons. (The reason we
never noticed before is none of the tests tried to _prepare_ a SELECT statement
with no keys; instead, they issued a query.)

Firstly, the handling of an empty set of parameters in `on_execute` was broken;
it created a `vec![vec![]]` instead of `vec![]`, causing it to hit a Vec1 assert
in `multi_lookup` and fail. This is now handled properly, and a new assert is
added for the case where columns have been rewritten to give a more descriptive
panic if that were to happen.

Secondly, bogokeys were excluded from the result set schema in `handle_select`,
but not in `prepare_select`, causing `do_read` to get very confused. This
oversight was corrected.

The second bug was rendered momentarily very annoying to debug because of the
occurrence of a double panic, caused by the `msql-srv` `RowWriter` panicking in
its Drop impl. It now throws away its `Result` instead of `unwrap()`ping it to
avoid this.