Add the start of a verify mode for noria-logictest
Add the beginnings of a verification runner for noria-logictest.
Currently this runs both statements and queries, and supports making
assertsions on the *number* of results from the query against the stored
number of results in a Digest result assertion. Included in this commit
is a very basic sqllogictest test script that runs successfully with
this mode.

Not yet implemented:
- Different sort modes
- Checking the digest against the returned results
- Checking the returned results against expected results in non-digest
  mode
- Hash threshold setting.

This is currently also mostly blocked because it appears our mysql
adapter is returning column values as the wrong type - they're coming
back as Bytes when they should be coming back as Ints! This could
however be me misunderstanding the mysql binary protocol.