Allow running logictests against mysql
Add a command-line flag (and corresponding run option fields) to run
sqllogictest tests against mysql instead of noria, for use as comparison
and verification. This also refactors the RAII stuff we're doing to
clean up after the test run to happen on a Drop impl of the RunOpts,
rather than a specific Deployment type, since it now has to also drop a
mysql db.