Implement sort modes, and compare types correctly
It turns out that it *is* in fact expected behavior for a compliant
mysql server to return a Bytes value for a column of type Integer -
mysql itself does this! So instead of *checking* the column types
against our expected columns, use them to *convert* column values to our
ColumnValue struct.

Now that we have proper column values, we can also implement sort modes,
using a custom Display impl per the sqllogictest docs and an Ord impl
based on comparing the resulting displayed strings.