Check md5 digests of results
Implement a function that calculates the md5 checksum of the results of
a query using the same method that sqllogictest does (with a newline
character in between each result) and use this to check the digest of
the results returned by queries against the db. I've run a reference
test from sqllogictest against mysql to confirm that this all works as
expected.