Try converting floats to ints
Some tests in sqllogictest have queries that return floats for a column,
but the tests themselves make assertions on ints. Regardless of whether
or not I think this is a good idea (I don't, for the record), we want to
support running arbitrary upstream sqllogictest tests unchanged, so
let's just try truncating all floats to integers in cases where we're
looking for an integer.