Move generally-useful utilities into Launchpad
Move some of the generally-useful (eg not crate- or domain-specific)
utilities we've been accumulating into the Launchpad crate, so that they
can be used across the workspace rather than only in the place where
they're defined. The most substantial part of this is the Bound- and
interval-related utilities, which have been moved from
unbounded-interval-tree into a new `intervals.rs` module in launchpad.
The only function that *didn't* move there is the `cmp` function for
comparing entire ranges, since the specific ordering that uses is highly
specific to the implementation of unbounded-interval-tree, and I don't
feel like it's going to be generally useful (though obviously it could
be extracted at some point if it turns out I'm wrong).

The main changes here in addition to things moving between modules are
the addition of more documentation and doctests for all top-level items,
since that's more important for utilities that are intended to be widely
used.