Refactor filter check logic and aggregate filters
Logic for checking whether a record passes a filter was duplicated in
many places. Now, this logic is consolidated in FilterVec tuple struct,
which contains a vector of filter conditions, and can be checked against
a given record to return a bool value.

Aggregates with filters (e.g. SUM(CASE WHEN ... THEN .. ELSE) was
originally a separate operator from normal aggregate. Refactor so that
aggregates just have an optional filter parameter and logic for
sum/count/avg not duplicated.

Note that there is still a clear distinction between Aggregate and
FilterAggregate in MIR, due to more complex optimization logic involving
filter aggregates. However, both aggregate and filteraggregate are
translated to the same Aggregate ingredient operator.