Actually test the other 2 crates in the `noria` dir
Commit 55332dbefd6d46bdd971ab9f9b1c7ccd495d01e2
("Use a Cargo workspace for the whole monorepo") inadvertently
exempted the `noria-server` and `noria-applications` crates from
automated Buildkite CI testing. This commit corrects that error.

Unfortunately, the tests now fail as a result of changes to the
repo since that commit. (They also didn't compile, so a small
fix in this commit makes them actually compile.)
This is pretty bad :(