Send SECURE_CONNECTION in the mysql handshake
The node.js mysql2 mysql client library doesn't have any internal
support for 4.1-protocol mysql connections that don't have the
SECURE_CONNECTION flag set - all code paths seem to assume that this
flag is set and not having it set gives an undefined value exception.
To make things a little bit easier to track, this commit defines a
constants.rs file with all of the various bit flags extracted from the
mysql source code itself, and uses those bitwise-or'ed together then
converted to a byte slice to get the actual capability flags to send to
the client as part of the handshake.