Add a test for a range query that partially misses
Add a test covering a range upquery that partially misses, and hence
needs to upquery for *part* of its missing state, which it then waits
for.

I added this test in an attempt to reproduce a bug elsewhere but was
unable to do so - this test seems generally nice to have though.