Only trigger recursive upqueries on misses
If we miss in an internal node when attempting to respond to an upquery,
we currently initiate a recursive upquery (on_replay_miss) to the
parents of said internal node. Previously, the code that did this
reproduced the key that was looked up directly - this was fine for point
queries, except in the case of range queries we can miss on only
a *subset* of the range we're looking into - we then need to initiate
multiple upqueries for the chunks of that subset.