Add support for reader-side ORDER BY / LIMIT, incl. slow ranged TopK
The `do_order_limit` function in noria/server/src/worker/readers.rs provides
a way to sort and limit an arbitrary iterator in the most zero-cost way
possible. To do this, we use the new `OrderedLimitedIter` type to avoid dynamic
dispatch (which is an enum of 4 `Iterator`s that simply forwards calls through).

We use this function to implement ORDER BY / LIMIT in the reader, by simply
wrapping the iterator of results the reader returns with a call to
`do_order_limit` before serializing the results. In order to pass through the
parameters (which column to order on, and which limit to use), queries against
a `View` now use the new `ViewQuery` struct instead of a 4-tuple (which
we attempt to use in as many places as we can, cleaning up things a
bit).

Importantly, this means we now have a very slow form of ranged TopK... because
we can retrieve a lot of rows from our ranged upquery and do the actual TopK
bit at the reader's end. Optimizing this is still a problem that needs to be
tackled; the TopK design doc contains some thoughts and suggestions.

There are still obstacles preventing this from actually working in all cases,
though. The query deduplicator bug PR #11 attempted to fix is a problem; we
include a very bad hack in the `NoriaConnector` that nulls out the ORDER BY
part of a SELECT before passing it on to noria proper iff there's no LIMIT as
well as a partial workaround.

Additionally, there are still bugs in ranged upqueries that have a TopK in the
replay path, as the ignored `design_doc_topk_with_preload` test shows (running
it causes noria to hang indefinitely, at least on my machine).

Two new tests, `order_by_basic` and `design_doc_topk`, are included to validate
the new behaviour.