Parse ILIKE and NOT ILIKE in nom-sql
Add ILIKE and NOT ILIKE enum members to the BinaryOperator type in
nom-sql, and add support for parsing both to the
binary_comparison_operator parser. To make this work properly (and not
be ambiguous with a column alias without AS) this also had to add ILIKE
as a reserved keyword.