Don't wrap strings with an extra layer of quotes
The ToString impl for DataType returns strings surrounded by
double-quotes - this was used when serializing result sets to mysql,
meaning that all string columns returned to mysql clients were wrapped
in an extra layer of double quotes. This changes that to instead use
Into<&str>, and covers it with a simple integration test that selects
strings out of a table.