Add aliases when queries are fully reused
Sometimes, we get a query that we can fully reuse the query graph from
another query for, but the actual *AST* is different. This can happen if
the queries are identical *after rewriting* but not before, for example
the following two queries:

    QUERY q1: SELECT * FROM t1 WHERE x != 1;
    QUERY q2: SELECT * FROM t1 WHERE NOT (x = 1);

Previously, we'd successfully reuse the query graph from q1 for q2, but
never be able to actually *look up* q2, because we forgot that they
could have the same reader! This changes the query reuse mechanism to
remember the *name* of the query we're fully reusing and communicate
that back to the Recipe by way of QueryFlowParts, and then the Recipe
can use that information to add a new alias for q2 pointing at q1.