Update server-components-demo to work with noria
Update (most of) the server-components demo to work via a connection to
the noria mysql adapter. This involves swapping out postgres for mysql2
as the database adapter (which requires writing a mysql version of the
postgres react suspense adapter, since there isn't an official mysql
react suspense adapter yet), and rewriting a bunch of the query syntax
to be mysql-compliant.

This doesn't fully work yet due to some feature gaps in noria, but the
seed script runs (it hangs, but if you ctrl+c it the data ends up in
noria) and the application can *connect* to noria (it immediately
crashes afterwards).