Avoid conflating the replay_key and the miss_key
Say we're doing an upquery for the KeyComparison `k`.
The `on_replay_miss` function on Domain is clever; it appreciates that,
when you're doing this upquery and you miss, the KeyComparison `m` that
you miss on might not be the same as `k`. This is why it has two
arguments: the `replay_key` is the KeyComparison you were originally
trying to replay (`k` in this case), and the `miss_key` is the miss you
had when you were doing that (`m` in this case). Whoever wrote this
function was clearly very forward thinking, which is nice!

Unfortunately, the new range query code failed to follow up on that
forward thinking mindset (quite understandably, since it wasn't entirely
obvious) and would clone & pass in the `miss_key` in as the `replay_key`.
This resulted in ranged upqueries hanging forever when parts of the
range were already materialized (i.e. when you'd miss on subsets of the
range and not the entire range), since the clever forward-thinking code
was fed garbage and never got a chance to do this magic.

This commit corrects that annoying oversight, meaning the
`design_doc_topk_with_preload` test now passes (which would exercise
that failure case in the previous paragraph). It also attempts to
comment what went wrong to make it clear what's happening with that bit
of the code, so hopefully this won't happen again!

(I would also like to humbly request that functions like
`on_replay_miss` actually have some doc comments added to them; this
seems like a case where that would have been useful)