Add a stub of the ParamFilter operator
Add a stub implementation of a new operator called ParamFilter, which
stores records matching a predicate on some value, keyed by that value.
There's a design doc, linked in the diff, for the details of this
operator, but the short explanation is that this will allow us to
implement table-scan filters like ILIKE in a way that can be better
integrated with the query graph, rather than doing them on the reader.

This is currently mostly unimplemented, and is here so we can build a
MIR implementation for it while the implementation is finished.