Make LikePattern construction never fail
Rather than just blindly accepting whatever string comes the way of the
LikePattern and then panicking if it's not a valid regular expression
post-transformation, actually escape regex tokens before passing them to
Regex::new, so that the `expect` in LikePattern::new is actually a
system invariant. I've tested this change with a proptest test and by
setting PROPTEST_CASES to 100000, so as a result I feel rather confident
that this will actually never fail.