Add a method to coerce DataTypes to another type
Add a `DataType::coerce_to` method, to coerce a DataType to another
target SqlType.

The semantics of SQL are that types get implicitly coerced on various
different operations - 1.0 and 1 compare as equal, strings get parsed as
timestamps, integers of different sizes compare as equal, etc. Since
noria *hashes* DataTypes when using them as lookup keys into a view, we
don't actually replicate this behavior - if you try to look up a row
with an integer primary key using a float value you'll get no results!
This is especially no good right now since JavaScript doesn't
actually *have* integers - so all numeric query parameters get passed as
floats!

The addition of this method lays the groundwork for across-the-board
casting of values - both keys on lookup and values for columns on
inserts and updates - to the type of their target column. There's a fair
bit of type coercion already present in this method, but I wouldn't be
surprised at all if down the road we have to add more type casts as we
discover more cases where the semantics of SQL expect types to get cast.