Coerce lookup keys to the type of the column
Use the schema of the getter for views, along with the new coerce_to
method on DataType, to coerce all lookup keys on reads to the types of
the column they're looking up. Right now, we need this because
JavaScript is incapable of sending integers for query parameters (since
it doesn't, y'know, *have* integers >.<) but generally this is something
that the semantics of SQL expect.

This version of the commit `unwrap`s any coercion errors, but soon we're
going to have a real error type on these methods which we can *return*
instead - there's a TODO here to that effect.