Coerce written values to the type of the column
Similar to what we're doing for lookup keys, use the schema of the table
to coerce written values, both in updates and inserts, to the type of
the column they're being written to. In this case, this allows us to
write timestamps by passing a properly-formatted string, which is
necessary for the react server-components-demo application.

There's a similar caveat here to the lookup keys - the coercion errors
are currently being `unwrap`ped rather than being returned as proper
error return values, since there's a separate change coming soon to add
an actual error enum to all these methods.