Add script to run the vote benchmark on Windtunnel
Add a mode to the vote benchmark to write benchmark results in a
windtunnel-compatible CSV format, and add a script to run that benchmark
on Windtunnel and report results.

Currently windtunnel just runs shell scripts directly, so we have to
manage setting up dependencies and shared volumes via docker containers
ourselves, but soon this is going to be replaced with a windtunnel.yml
config or something like that.