In SqlToMirConverter::add_leaf_below, specify new node as 'node'.
Typically MirNodeType::Leaf's node parameter is its parent. In add_leaf_below,
set the new Leaf's node to be the leaf's parent, rather than its grandparent.
The MirNodeType::Leaf node parameter does not appear to be used in dataflow, so
this change is for consistency with other usage of MirNodeType::Leaf rather than
a functional change.