Add prep+parametrized query support to logictest
Add support for running prepared, parametrized queries with
noria-logictest, so that we can exercise the various failure scenarios
with noria that arise especially from parametrized queries. The syntax,
which could easily be bikeshedded, here looks like:

    query ITT nosort
    SELECT * FROM notes WHERE id = ?;
    ? = 1
    ----
    1
    test_1
    test_1

Or, for numbered instead of named params:

    query ITT nosort
    SELECT * FROM notes WHERE id = $1;
    $1 = 1
    ----
    1
    test_1
    test_1

Any number of those ? = whatever or $n = whatever lines can be specified
after the query but before the line separating the query and the
expected results.