Always ensure the `target` docker volume exists
Add a buildkite pre-command hook to create the `target` docker volume if
it doesn't already exist. This is necessary if one of the steps that
uses docker-compose (noria-mysql tests or logictest) is the first step
to run on a brand new (recently scaled-up) agent. Historically we just
got away with this because all the other steps that use just docker
would have created the volume anyway.