Possible fix to coercion from text to date
Currently, there is no way to coerce any type
(specifically, Text/TinyText/MediumText/Varchar) into a Date.
This behavior breaks loading the test_db database.
This workaround adds the match-clause to allow
coercing a text type into a Date,
which will internally transform the text into a NaiveDate,
and then map it into a Timestamp.
The NaiveDateTime passed to the Timestamp
is just the NaiveDate with a NaiveTime::from_hms(0, 0, 0).