domain: Deduplicate ranged upqueries
An important invariant in Noria is that we never have more than one
active upquery for the same key from the same reader along the same
replay path - lots of stuff internal to noria assumes that this is the
case. Previously, the deduplication of ranged upqueries was done
per-KeyComparison, which worked fine for point queries and *fully*
deduplicated range queries, but allowed us to still upquery for
multiple overlapping ranges, violating the invariant.

This changes the `reader_triggered` data structure in the Domain from a
HashSet of KeyComparison to an IntervalTree, which allows us to take all
upqueries and *subtract* the set of existing pending upqueries from
them, and then only actually upquery for the subset of the range that we
are not already waiting on.

Change-Id: If634cb023f3d6a241da4e499ae36f4b5fbc5621c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/804
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>