Implement MySQL password authentication
A specific deployment scenario requires us to deploy the MySQL adapter
open to the internet so that a non-ip-stable serverless environment can
connect to it. To make this secure, this commit implements MySQL's
"secure password" authentication scheme in msql-srv. This avoids the
complexity of TLS or SSL, while still getting us authentication without
sending plaintext passwords over the wire: the server sends 20 random
bytes of data with the initial handshake, and the client hashes that
concatenated with the password for the response.

The actual method by which `MysqlShim`s indicate to msql-srv what the
password for a user is is a little half-baked here - essentially I've
just added a new `password_for_username` method that takes a username
and returns an optional password. noria-mysql now takes two new command
line arguments, an optional username (defaulting to "root") and a
required password, and uses those as the one and only pair of
authentication credentials for all clients. This has a few problems - we
only ever support one pair of credentials, for one, and for another
having the credentials passed on the argv is pretty insecure (processes
running on the same machine can trivially read each others' argv), but
for now this solves the immediate problem of wanting to authenticate
noria-mysql running over the public internet, and we can improve on the
remaining bits later.

I've tested this both with a new integration test in msql-srv itself
using the *rust* mysql client library, and via connecting to noria-mysql
via the nodejs mysql client library, and in both cases connections were
successful when using the correct password, and failed in the expected
way when using the incorrect password.