Make noria-mysql and msql-srv asynchronous
The `msql-srv` MySQL server crate is now fully asynchronous. This is done by
reworking the core packet writer and reader in `src/packet.rs` to be async, and
then propagating the change up the tree (a lot of s/fn/async fn/ etc.)

Annoyingly, the crate relies on `mysql_common` to serialize MySQL protocol,
which only works on standard `std::io::Write` writers. We temporarily hack
around this by writing the protocol data into an in-memory buffer, and then
writing the buffer; this is an area for future optimization.

As part of this, `noria-mysql` is now fully async too -- we no longer block on
either the Noria side or the MySQL client side, which should improve things!
The MySQL backend connector code was hackily ported, and could do with more
work before being put into production.