Support for timestamps in evbtree.
This change enables timestamps to be set/get in the evbtree.
This will enable noria reader nodes to maintain a timestamp in the
evbtree and will allow reader workers to block reads on a timestamp
ordering in a future commit. These timestamps will be used to implement
read-your-write consistency by blocking reads on the value of the
evbtree timestamp.

The implementation follows the support for metadata in the evbtree.
A new value is stored in the evbtree `timestamp` that can be set and
is only readable following a publish.