Require that subquery views be named as specified.
Ensure that subquery views are created with the name actually requested, and not
with an alternative name when a preexisting query is reused. This is necessary
to ensure that existing column references can refer to the subquery correctly
and to ensure that distinct subqueries are treated as separate views in the
query graph even if they are "exact matches" in the sense of
QueryGraphReuse::ExactMatch. Adds a new 'is_name_required' argument to
SqlIncorporator::nodes_for_named_query but does not modify SqlIncorporator’s
public interface.