RYW Consistency: View queries can specify timestamps.
ViewQuery now has a timestamp value that is compared against
the timestamp at the evbtree by the reader. This allows queries
to specify that (1) they want ryw consistency, and (2) that the
data in the evbtree map returned meets their consistency requirements.

As the dataflow currently does not propagate a timestamp to the reader
node, and subsequently the evbtree map, this is currently a no-op.