RYW consistency: Refactor Table::Input to a generic packet data type.
The existing flow to interact with base tables is to (1) get a handle
to the base table, (2) create a Table::Input struct with any table
operations to be performed, (3) pass that to the replica which
converts this into an Input packet and pass this through to the
base table data flow node. With read-your-write consistency we will
now also pass a timestamp to the base table that will be converted
into a timestamp packet.

This change creates the `PacketData` type which represents any type
of data that will be converted into a packet, namely `Timestamp` and
`Input`. Packets that previously wrapped Input, will now wrap
PacketData. Future Timestamp packets will be able to wrap PacketData
that holds relevant timestamp information.