RYW consistency: Timestamp skeleton and comparison support.
This introduces three new Timestamp function skeletons `new`,
`min`, and `join`, required by the server and client libraries.
Future changes will be responsible for the implementation. It also
adds the map from base table to timestamp and implements `satisfies`,
a function that evaluates if a timestamp can satisfy the
read-your-write consistency requirements of another.