Insert unique views when table alias removal would create ambiguous references.
Analyze a SqlQuery's table aliases to identify cases where alias removal would
create ambiguity. In those cases, replace the aliased table reference with a
reference to a new globally unique view (in the same manner that views are
inserted for explicit subqueries). Handling of unambiguous aliases remains
unchanged (they are simply rewritten to refer to the underlying table). Mir and
dataflow graphs continue to refer to distinct tables and remain unaware of table
aliases.