Fixed debezium connector not handling MariaDB's dummy events correctly
This commit modifies the `debezium-connector-application` and makes the `Source
.table` an `Option<String>`, versus a simple `String`, skipping Kafka messages
that have a `None` table. The reason for this change is that MariaDB sends a
dummy event message to the [database.name.server] topic, which has a `null`
value for most of the fields, which then makes the Debezium Connector
Application to fail at parsing it, bringing the whole application down. With
this, we can ignore those messages.