Minor improvement to the view_checker.rs tool
This commit adds the capability to specify which view query to run against Noria
 as a command line parameter, thus avoiding having to recompile the test
each time a new query needs to be tested.
The new command line parameter is `--query [name of the query]` or `-q [name of
the query]` for short. If no query is specified, a default value will be used,
which is `testQuery` as it was hardcoded before, for backwards compatibility.